requirements.txt
streamlit
langchain
langchain-chroma
langchain-huggingface
langchain-groq
langchain-community
python-dotenv
chromadb
sentence-transformers
torch

